refactor(doc): split doc.rs (1917 lines) into 3 focused modules
- extract.rs: data model (DocEntry, DocKind, etc.) and AST walking
- render.rs: markdown, JSON schema, and coverage rendering
- prelude.rs: multi-file categorised prelude reference generation
- mod.rs: entry points and public API re-exports

Additional improvements:
- Add FileCouldNotBeWritten error variant for write failures
  (previously misused FileCouldNotBeRead)
- Refactor extract_see_also to reduce nesting depth from 10 to 4
  via helper functions (extract_see_also_from_block, collect_symbol_refs)
- Replace pointer-as-hash deduplication in generate_category_page
  with index-based approach using entry_index helper
- Fix json_str fallback to use serde_json::Value instead of
  incomplete manual escaping
- Refactor extract_block_str and extract_block_sym to use early
  returns instead of deep nesting
